python/examples/example_website/requirements.txt
channels~=4.0.0
channels-redis~=4.1.0
daphne~=4.0.0
django_plotly_dash~=2.1.4
django-redis~=5.2.0
dpd-static-support~=0.0.5
plotly~=5.14.1
numpy~=1.24.3
pandas~=2.0.1
redis~=4.5.4
tradingeconomics~=4.2.3
python-dotenv~=1.0.0
whitenoise~=6.4.0
